+++ apps/docs/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @aparte-workspace/docs
 
+## 0.0.18
+
+### Patch Changes
+
+- Updated dependencies [705224b]
+- Updated dependencies [9df343c]
+- Updated dependencies [39b777f]
+  - @aparte/core@0.16.2
+  - @aparte/locale-fr@0.16.2
+  - @aparte/plugin-shiki@0.16.2
+
 ## 0.0.17
 
 ### Patch Changes